Create a set of Font Awesome or Bootstrap Markers

This function is intended as a more up-to-date implementation of leaflet::awesomeIcons(). Key benefits include having access to modern Font Awesome and Bootstrap icons, and allowing any colour to be used the marker and icon.

Usage

magicIcons(
  icon = "circle",
  markerColor = NULL,
  iconColor = NULL,
  markerSize = 30L,
  library = "fontawesome",
  className = NULL
)

Arguments

icon

Name of the Font Awesome icon, passed to fontawesome::fa() or bsicons::bs_icon(). A full list of available icons can be found using fontawesome::fa_metadata() or at https://icons.getbootstrap.com/.

markerColor

The color of the teardrop-shaped marker.

iconColor

The color of the fontawesome icon.

markerSize

The size of the marker. Defaults to 30, which is roughly the same size as leaflet::addMarkers().

library

One of "fontawesome" or "bootstrap", defining the icon library of interest. Defaults to "fontawesome".

className

a custom class name to assign to both icon and shadow images

Value

a leaflet::iconList(), to be passed to the icon argument of leaflet::addMarkers()

Author

Jack Davison

Examples

# create map
library(leaflet)
port_talbot %>%
  dplyr::mutate(
    icon = dplyr::case_match(
      site_type,
      "Urban Industrial" ~ "industry",
      "Urban Traffic" ~ "car",
      "Urban Background" ~ "city",
      .default = "x"
    ),
    color = dplyr::case_match(
      site_type,
      "Urban Industrial" ~ "#12436D",
      "Urban Traffic" ~ "#801650",
      "Urban Background" ~ "#28A197",
      .default = "#3D3D3DFF"
    )
  ) %>%
  leaflet() %>%
  addProviderTiles("CartoDB.Voyager") %>%
  addMarkers(
    icon = ~ magicIcons(icon, color, "white"),
    popup = ~site
  )
#>  ■■■■■■■■■■■                       33% |  ETA:  3s
#> Assuming "longitude" and "latitude" are longitude and latitude, respectively